🌍mapper function now returns any
